cbuf: use size_t instead of unsigned long

+++ cbuf.c
@@ -27,16 +27,17 @@
  */
 
 #include <assert.h>
+#include <stddef.h>
 
 #include "cbuf.h"
 #include "error.h"
 #include "macros.h"
 
 /* Negative close to 0 so that an overflow occurs early */
-#define CBUF_INIT_INDEX ((unsigned long)-500)
+#define CBUF_INIT_INDEX ((size_t)-500)
 
 void
-cbuf_init(struct cbuf *cbuf, char *buf, unsigned long capacity)
+cbuf_init(struct cbuf *cbuf, char *buf, size_t capacity)
 {
     assert(ISP2(capacity));
 
@@ -46,8 +47,8 @@ cbuf_init(struct cbuf *cbuf, char *buf, unsigned long capacity)
     cbuf->end = cbuf->start;
 }
 
-static unsigned long
-cbuf_index(const struct cbuf *cbuf, unsigned long abs_index)
+static size_t
+cbuf_index(const struct cbuf *cbuf, size_t abs_index)
 {
     return abs_index & (cbuf->capacity - 1);
 }
@@ -77,7 +78,7 @@ cbuf_pop(struct cbuf *cbuf, char *bytep)
 }
 
 int
-cbuf_read(const struct cbuf *cbuf, unsigned long index, char *bytep)
+cbuf_read(const struct cbuf *cbuf, size_t index, char *bytep)
 {
     /* Mind integer overflows */
     if ((cbuf->end - index - 1) >= cbuf_size(cbuf)) {
